### 2016-05-27 #24 @whedon
Submission: corner.py: Scatterplot matrices in Python
**Submitting author:** @dfm (<a href="http://orcid.org/0000-0002-9328-5652">Daniel Foreman-Mackey</a>)
**Repository:** <a href="https://github.com/dfm/corner.py" target ="_blank">https://github.com/dfm/corner.py</a>
**Version:** v2.0.0
**Archive:** <a href="http://dx.doi.org/10.5281/zenodo.53155" target ="_blank">http://dx.doi.org/10.5281/zenodo.53155</a>

- [x] **Archive:** Does the <a target="_blank" href="http://dx.doi.org/10.5281/zenodo.53155">software archive</a> resolve?

- [x] **Performance:** Have the performance claims of the software been confirmed?

- [x] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?
