### 2016-05-28 #24 @tacaswell
I actually have 1 commit into this project (completely forgot about that until my name showed up in the author list on zenodo).  It is an almost trivial change: https://github.com/dfm/corner.py/commit/8b87b92739974ce48f84f0b43e2c208472b8035f

I suspect that I am being pedantic about conflict of interest here, but is this a problem?
