### 2016-05-28 #24 @arfon
Thanks for bringing this to my attention. Given that the commit is very small I think it's ok to proceed with the review.
