### 2016-05-28 #24 @genomematt
@arfon I think for COI we should say bug fix patches are ok (need to be declared) but not feature additions.  This is clearly a bug fix.  With reviewers who are active and projects that are useful bug fixes may be common, and the last thing we want is to inhibit our reviewers from fixing bugs.
