### 2016-05-29 #24 @arfon
@genomematt - I've added a small clarification here: https://github.com/openjournals/joss/commit/dad06058f66a9681508c907fb44b27f43e37c2c9

Perhaps we should add something to the reviewer guidelines too?
