### 2016-06-08 #24 @tacaswell
@dfm `quantile` is exported in the `__all__`, but is not in the built docs nor is it's docstring numpydoc format.  Was this intended to be part of the public API?

This paper makes no performance claims, how does that get handled?

The mpl paper has a DOI http://dx.doi.org/10.1109/MCSE.2007.55

The citation to the ADS search results seems strange to me, I think it is ok to assert an 'also known as' with out citation.
