### 2016-06-08 #24 @tacaswell
I recommend publication on revision of my minor comments.
