### 2016-06-08 #24 @arfon
> This paper makes no performance claims, how does that get handled?

In that case you can just check the box.
