### 2016-06-08 1c61cf3
adding tests and docstring for quantile
