### 2016-06-08 1029e1e
adding quantile to API docs
