+++ docs/api.rst
@@ -4,7 +4,9 @@ Detailed API documentation
 ==========================
 
 The main user-facing function is :func:`corner.corner` but the lower level
-function :func:`corner.hist2d` is also documented here.
+functions :func:`corner.hist2d` and :func:`corner.quantile` are also
+documented here.
 
 .. autofunction:: corner.corner
 .. autofunction:: corner.hist2d
+.. autofunction:: corner.quantile
