### 2016-06-08 #24 @dfm
@tacaswell: Thanks for the suggestions. I've added some tests and documentation for the quantiles function. I added the DOI for the matplotlib paper and removed the ADS citation.

Thanks!
