### 2016-06-08 #24 @arfon
Thanks for the review @tacaswell.

@dfm your DOI will be http://dx.doi.org/10.21105/joss.00024 once the Crossref queue is processed (looks a little backed up right now).
