### 2016-06-08 #24 @dfm
@arfon: what are your thoughts on submitting a paper from joss to arxiv?
On Wed, Jun 8, 2016 at 12:05 PM Arfon Smith <notifications@github.com>
wrote:

> Closed #24 <https://github.com/openjournals/joss-reviews/issues/24>.
>
