### 2016-06-08 #24 @arfon
> @arfon: what are your thoughts on submitting a paper from joss to arxiv?

Fine by me. My only request would be that you update the arXiv metadata to point to the JOSS DOI.
