+++ paper/paper.bib
@@ -5,7 +5,8 @@
     Pages = {90--95},
     Title = {Matplotlib: A 2D graphics environment},
     Volume = {9},
-    Year = {2007}
+    Year = {2007},
+    doi = {10.1109/MCSE.2007.55}
 }
 
 @online{corner_github,
@@ -13,14 +14,6 @@
   title = {corner.py on GitHub},
   year = 2016,
   url = {https://github.com/dfm/corner.py},
-  urldate = {2016-05-26}
-}
-
-@online{adsabs,
-  author = {Astrophysics Data System},
-  title = {ADS Search Results},
-  year = 2016,
-  url = {https://ui.adsabs.harvard.edu/},
   urldate = {2016-05-26}
 }
 
