+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{matplotlib,
-    Author = {Hunter, John D and others},
+    Author = {Hunter, John D},
     Journal = {Computing in science and engineering},
     Number = {3},
     Pages = {90--95},
@@ -21,5 +21,6 @@
   author = {{Zenodo Archive}},
   title = {corner.py: Scatterplot matrices in Python},
   doi = {10.5281/zenodo.53155},
+  year = 2016,
   howpublished = {\url{http://dx.doi.org/10.5281/zenodo.53155}}
 }
