### 2016-06-08 #24 @dfm
Done. One request for the future: I had to download the pandoc latex template and generate the .tex file to submit to arxiv (they spot PDFs that were generated by LaTeX) so it would be great to make the tex available somewhere!
