### 2016-06-08 #24 @danielskatz
I notice that "References: Do all archival references that should have a DOI list one (e.g. papers, datasets, software)?" never got checked in the review.  There are references listed, and while one has a DOI, the Matplotlib paper also should (specifically, 10.1109/MCSE.2007.55).  I wonder if we have a process problem here?  It seems like the review shouldn't be finalized and the paper shouldn't be published until all the boxes are checked.  Also for the third reference, I don't think Astrophysics Data System is a proper name that should be written as "System, Astrophysics Data".
