### 2016-06-08 #24 @dfm
These changes were made here: https://github.com/dfm/corner.py/blob/master/paper/paper.bib ... I'm not sure if there was something else I needed to do.
