### 2016-06-08 #24 @danielskatz
It looks like there was a paper regeneration step that didn't get done.  @arfon ?
