### 2016-06-08 #24 @arfon
> It looks like there was a paper regeneration step that didn't get done. @arfon ?

Darn it, yes, looks like human error (on my part). Will update.
