### 2016-06-09 #24 @dfm
It looks like you have updated it now. Should we close this issue again or are there outstanding problems?
